requirements.txt
attr==0.3.1
BenchExec==2.7
certifi==2020.6.20
chardet==3.0.4
Deprecated==1.2.10
future==0.18.2
idna==2.10
pipfile==0.0.2
plumbum==1.6.9
python-decouple==3.3
PyVimeo==1.1.0
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
Tempita==0.5.2
tinydb==4.1.1
toml==0.10.1
tuspy==0.2.4
urllib3==1.25.10
virtualenv==16.7.10
wrapt==1.12.1
